backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
grpcio
grpcio-tools
apscheduler
pyotp